requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

sqlmodel==0.0.22
sqlalchemy==2.0.25

psycopg2-binary==2.9.9

alembic==1.13.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0

python-dotenv==1.0.1

# New dependencies
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.21
